+++ src/main/java/com/microtech/smartshop/service/ClientStats.java
@@ -2,8 +2,10 @@ package com.microtech.smartshop.service;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ClientStats {
     private Integer totalCommandes;
-    private Double totalDepense;
+    private BigDecimal totalDepense;
 }
